+++ data/productservice/src/main/java/com/jlndev/productservice/data/repository/model/ProductItemModel.kt
@@ -6,5 +6,5 @@ data class ProductItemModel(
     val description: String,
     val price: Double,
     val image: String,
-    var quantity: Double = 1.0
+    var quantity: Long = 1
 )
